test(dot prod): New test for both muladd and add
The test performs the dot product between 2 couples of vectors, in which
one is shared among the vectors. This allows the compiler to spot a SIMD
situation easily.

If one provides `USE_BOOL_INPUTS=1` with the TCL script, two vectors are
boolean, while one isn't. This emulates what happens in an SNN, main
target of the simd add optimization.