fix(SILVIAMuladd): avoid arguments extension with unknown signedness
Before this change, the unextended arguments were sign-extended to 8 bits.
However, this may leads to broken code since the argument may require
zero-extension.

This change avoids extension of unknown-signedness arguments by truncating the
extended arguments.
If the argument is originally (i.e., without any extension) shorter than 8
bits, it means it is the case of a `mul i7` (or less) operation, where the
higher bits are discarded. Thus, it is safe to zero-extend its arguments.