fix(SILVIAMuladd): include a guard MSB to the _simd_muladd_2 arguments
The arguments of _simd_muladd_2 are sign-extended.
This change includes an additional guard bit to the arguments.
This additional bit is used for the sign-extension: if the argument is signed,
this additional bit carries the sign; if the argument is unsigned, this
additional bit is 0 and preserves the functionality despite the sign-extension
on the unsigned argument.